+++ src/Monitor/Monitor/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
         
 ## Upcoming Release
+* A new parameter called ActionGroupId was added to New-AzMetricAlertRuleV2 command.
+	- The user can provide either ActionGroupId(string) or ActionGorup(ActivityLogAlertActionGroup).
 
 ## Version 1.5.0
 * Update references in .psd1 to use relative path